@@ -1,161 +1,91 @@
 package com.perfree.model;
 
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import org.hibernate.validator.constraints.Length;
+import com.baomidou.mybatisplus.annotation.TableName;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
+import java.io.Serial;
 import java.io.Serializable;
-import java.util.Date;
-import java.util.List;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.Setter;
 
 /**
- * menu table
- * @author Perfree
+ * <p>
+ * 
+ * </p>
+ *
+ * @author perfree
+ * @since 2023-09-27
  */
-@ApiModel(value="Menu-菜单数据",description="菜单数据")
+@Getter
+@Setter
+@TableName("p_menu")
 public class Menu implements Serializable {
-    private static final long serialVersionUID = 7817277417501762477L;
-    public static int TYPE_FRONT = 0;
-    public static int TYPE_AFTER = 1;
 
-    @ApiModelProperty(value="菜单ID",name="id")
+    @Serial
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * 主键
+     */
     private String id;
 
-    @ApiModelProperty(value="父级菜单ID",name="pid")
+    /**
+     * 父级id
+     */
     private String pid;
 
-    @ApiModelProperty(value="菜单名",name="name")
-    @NotBlank(message = "菜单名不允许为空")
-    @Length(max = 30,message = "菜单名最多30个字符")
+    /**
+     * 菜单名
+     */
     private String name;
 
-    @ApiModelProperty(value="菜单url",name="url")
-    @NotBlank(message = "菜单链接不允许为空")
-    @Length(max = 50,message = "菜单链接最多50个字符")
+    /**
+     * 菜单链接
+     */
     private String url;
 
-    @ApiModelProperty(value="菜单图标",name="icon")
+    /**
+     * 菜单图标
+     */
     private String icon;
 
-    @ApiModelProperty(value="菜单序号",name="seq")
+    /**
+     * 排序序号
+     */
     private Integer seq;
 
-    @ApiModelProperty(value="菜单类型",name="type", example = "0:前台,1:后台")
+    /**
+     * 菜单类型0:前台,1:后台
+     */
     private Integer type;
 
-    @ApiModelProperty(value="菜单状态",name="status", example = "0:启用,1禁用")
-    @NotNull(message = "菜单状态不允许为空")
-    private Integer status;
-
-    @ApiModelProperty(value="创建时间",name="createTime")
-    private Date createTime;
-
-    @ApiModelProperty(value="更新时间",name="updateTime")
-    private Date updateTime;
-
-    @ApiModelProperty(value="菜单打开方式",name="target", example = "0本页,1:新窗口")
-    @NotNull(message = "打开方式不允许为空")
+    /**
+     * 菜单打开方式:0本页,1:新窗口
+     */
     private Integer target;
 
-    @ApiModelProperty(value="子菜单",name="childMenu")
-    private List<Menu> childMenu;
+    /**
+     * 菜单状态0:启用,1禁用
+     */
+    private Integer status;
 
+    /**
+     * 插件id
+     */
+    private String pluginId;
 
-    public List<Menu> getChildMenu() {
-        return childMenu;
-    }
+    /**
+     * 菜单标识:0:系统自带,1:用户创建,2:插件
+     */
+    private Integer flag;
 
-    public void setChildMenu(List<Menu> childMenu) {
-        this.childMenu = childMenu;
-    }
+    /**
+     * 创建时间
+     */
+    private LocalDateTime createTime;
 
-    public Integer getTarget() {
-        return target;
-    }
-
-    public void setTarget(Integer target) {
-        this.target = target;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getIcon() {
-        return icon;
-    }
-
-    public void setIcon(String icon) {
-        this.icon = icon;
-    }
-
-    public Integer getSeq() {
-        return seq;
-    }
-
-    public void setSeq(Integer seq) {
-        this.seq = seq;
-    }
-
-    public Integer getType() {
-        return type;
-    }
-
-    public void setType(Integer type) {
-        this.type = type;
-    }
-
-    public Integer getStatus() {
-        return status;
-    }
-
-    public void setStatus(Integer status) {
-        this.status = status;
-    }
-
-    public Date getCreateTime() {
-        return createTime;
-    }
-
-    public void setCreateTime(Date createTime) {
-        this.createTime = createTime;
-    }
-
-    public Date getUpdateTime() {
-        return updateTime;
-    }
-
-    public void setUpdateTime(Date updateTime) {
-        this.updateTime = updateTime;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getPid() {
-        return pid;
-    }
-
-    public void setPid(String pid) {
-        this.pid = pid;
-    }
+    /**
+     * 更新时间
+     */
+    private LocalDateTime updateTime;
 }
