@@ -1,115 +1,74 @@
 package com.perfree.model;
 
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
 
+import java.io.Serial;
 import java.io.Serializable;
-import java.util.Date;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.Setter;
 
 /**
- * @description Plugin model
- * @author Perfree
- * @date 2021/8/13 13:42
+ * <p>
+ * 
+ * </p>
+ *
+ * @author perfree
+ * @since 2023-09-27
  */
-@ApiModel(value="Plugin-插件数据",description="插件数据")
+@Getter
+@Setter
+@TableName("p_plugin")
 public class Plugin implements Serializable {
-    private static final long serialVersionUID = 7817277417501722377L;
-    @ApiModelProperty(value="插件ID",name="id")
-    private Long id;
 
-    @ApiModelProperty(value="插件名",name="name")
+    @Serial
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * 主键
+     */
+    @TableId(value = "id", type = IdType.AUTO)
+    private Integer id;
+
+    /**
+     * 插件名
+     */
     private String name;
 
-    @ApiModelProperty(value="插件路径",name="path")
+    /**
+     * 路径
+     */
     private String path;
 
-    @ApiModelProperty(value="插件描述",name="desc")
+    /**
+     * 插件描述
+     */
     private String desc;
 
-    @ApiModelProperty(value="插件版本",name="version")
+    /**
+     * 版本
+     */
     private String version;
 
-    @ApiModelProperty(value="插件作者",name="author")
+    /**
+     * 作者
+     */
     private String author;
 
-    @ApiModelProperty(value="创建时间",name="createTime")
-    private Date createTime;
+    /**
+     * 插件状态:0禁用,1启用
+     */
+    private Integer status;
 
-    @ApiModelProperty(value="更新时间",name="updateTime")
-    private Date updateTime;
+    /**
+     * 创建时间
+     */
+    private LocalDateTime createTime;
 
-    @ApiModelProperty(value="插件状态",name="status")
-    private Integer status = 0;
-
-    public Integer getStatus() {
-        return status;
-    }
-
-    public void setStatus(Integer status) {
-        this.status = status;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getPath() {
-        return path;
-    }
-
-    public void setPath(String path) {
-        this.path = path;
-    }
-
-    public String getDesc() {
-        return desc;
-    }
-
-    public void setDesc(String desc) {
-        this.desc = desc;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public Date getCreateTime() {
-        return createTime;
-    }
-
-    public void setCreateTime(Date createTime) {
-        this.createTime = createTime;
-    }
-
-    public Date getUpdateTime() {
-        return updateTime;
-    }
-
-    public void setUpdateTime(Date updateTime) {
-        this.updateTime = updateTime;
-    }
+    /**
+     * 更新时间
+     */
+    private LocalDateTime updateTime;
 }
