@@ -1,94 +1,53 @@
 package com.perfree.model;
 
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
 
+import java.io.Serial;
 import java.io.Serializable;
-import java.util.Date;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import lombok.Setter;
 
 /**
- * role table
- * @author Perfree
+ * <p>
+ * 
+ * </p>
+ *
+ * @author perfree
+ * @since 2023-09-27
  */
-@ApiModel(value="Role-角色数据",description="角色数据")
+@Getter
+@Setter
+@TableName("p_role")
 public class Role implements Serializable {
-    private static final long serialVersionUID = -7909060114936447269L;
 
-    @ApiModelProperty(value="角色ID",name="id")
-    private Long id;
+    @Serial
+    private static final long serialVersionUID = 1L;
 
-    @ApiModelProperty(value="角色名",name="name")
+    /**
+     * 主键
+     */
+    @TableId(value = "id", type = IdType.AUTO)
+    private Integer id;
+
+    /**
+     * 角色名
+     */
     private String name;
 
-    @ApiModelProperty(value="角色描述",name="description")
+    /**
+     * 角色描述
+     */
     private String description;
 
-    @ApiModelProperty(value="角色代码",name="code")
+    /**
+     * 角色编码
+     */
     private String code;
 
-    @ApiModelProperty(value="创建时间",name="createTime")
-    private Date createTime;
+    private LocalDateTime createTime;
 
-    @ApiModelProperty(value="更新时间",name="updateTime")
-    private Date updateTime;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public Date getCreateTime() {
-        return createTime;
-    }
-
-    public void setCreateTime(Date createTime) {
-        this.createTime = createTime;
-    }
-
-    public Date getUpdateTime() {
-        return updateTime;
-    }
-
-    public void setUpdateTime(Date updateTime) {
-        this.updateTime = updateTime;
-    }
-
-    @Override
-    public String toString() {
-        return "Role{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", description='" + description + '\'' +
-                ", code='" + code + '\'' +
-                ", createTime=" + createTime +
-                ", updateTime=" + updateTime +
-                '}';
-    }
+    private LocalDateTime updateTime;
 }
